fix: replace magefile with lint script
Trying to run magefile locally returns an error:

```
Installing Deps...
Updating...
Error: exec: "go-licenser": executable file not found in $PATH
exit status 1
```

I suspect this is because the `go get` syntax was change in Go 1.16
and it cannot be used to install packages in recent versions of go.

See https://go.dev/doc/go1.16#modules

> go get should be used with the -d flag to adjust the current module's
dependencies without building packages, and use of go get to build and
install packages is deprecated. In a future release, the -d flag will
always be enabled.

We could fix the magefile but it is polluting the dependency graph of
other projects (aws lambda extension), so I opted to replace the magefile
with a POSIX compliant shell script and remove the mage dependency.