Use RFC3339 in UTC as the default time format for log messages (#84)
* Use RFC3339 in UTC as the default time format for log messages

This ensures that times that are logged in messages can be deserialised
by the encoding/json package irrespective of the timezone that the time
is in when logged; currently if the time is not in UTC, the ISO8601
format that is used is not accepted. Also provide a non-UTC formatter
that can be used where the timezone is important.

* Add test for non-UTC timestamp

Also fix time validation in logger tests.

---------

Co-authored-by: Andrew Wilkins <axw@elastic.co>